Cleaned up the experimental bits of code.  Removed all packages to fit in github.
The first check-in of code worked,  I overran in time because I tried to retrieve json data using ajax,  but browsers block that because it is on a different domain from the running program.  Then I had to learn about git hub.  That took over an hour,  Then I hit size issues with github, and ended up removing all packages.  Just do a restore packages.